requirements.txt
aiogram
pymongo
python-dotenv
cyrtranslit